refactor: 랭킹 산출 쿼리 리팩토링
- 네이티브 쿼리 명시
- 맵퍼 클래스 수정
- 테스트및 검증 완료

+++ src/test/java/com/somemore/domains/volunteerrecord/repository/VolunteerRecordRepositoryTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -39,10 +40,10 @@ class VolunteerRecordRepositoryTest extends IntegrationTestSupport {
         assertEquals(5, results.size());
         
         // 각 등수별 인원 확인
-        assertEquals(100L, results.get(0)[1]);
-        assertEquals(100L, results.get(1)[1]);
-        assertEquals(90L, results.get(2)[1]);
-        assertEquals(60L, results.get(3)[1]);
+        assertEquals(100L, ((BigDecimal) results.get(0)[1]).longValue());
+        assertEquals(100L, ((BigDecimal) results.get(1)[1]).longValue());
+        assertEquals(90L, ((BigDecimal) results.get(2)[1]).longValue());
+        assertEquals(60L, ((BigDecimal) results.get(3)[1]).longValue());
     }
 
     @DisplayName("주간 봉사 시간으로 4위까지의 랭킹을 반환할 수 있다. 동점자를 같이 반환한다. (Repository)")
@@ -71,11 +72,12 @@ class VolunteerRecordRepositoryTest extends IntegrationTestSupport {
         assertEquals(5, results.size());
 
         // 각 등수별 인원 확인
-        assertEquals(100L, results.get(0)[1]);
-        assertEquals(100L, results.get(1)[1]);
-        assertEquals(90L, results.get(2)[1]);
-        assertEquals(60L, results.get(3)[1]);
-        assertEquals(50L, results.get(4)[1]);
+        assertEquals(100L, ((BigDecimal) results.get(0)[1]).longValue());
+        assertEquals(100L, ((BigDecimal) results.get(1)[1]).longValue());
+        assertEquals(90L, ((BigDecimal) results.get(2)[1]).longValue());
+        assertEquals(60L, ((BigDecimal) results.get(3)[1]).longValue());
+        assertEquals(50L, ((BigDecimal) results.get(4)[1]).longValue());
+
     }
 
     @DisplayName("월간 봉사 시간으로 4위까지의 랭킹을 반환할 수 있다. 동점자를 같이 반환한다. (Repository)")
@@ -104,10 +106,10 @@ class VolunteerRecordRepositoryTest extends IntegrationTestSupport {
         assertEquals(5, results.size());
 
         // 각 등수별 인원 확인
-        assertEquals(100L, results.get(0)[1]);
-        assertEquals(100L, results.get(1)[1]);
-        assertEquals(90L, results.get(2)[1]);
-        assertEquals(60L, results.get(3)[1]);
-        assertEquals(50L, results.get(4)[1]);
+        assertEquals(100L, ((BigDecimal) results.get(0)[1]).longValue());
+        assertEquals(100L, ((BigDecimal) results.get(1)[1]).longValue());
+        assertEquals(90L, ((BigDecimal) results.get(2)[1]).longValue());
+        assertEquals(60L, ((BigDecimal) results.get(3)[1]).longValue());
+        assertEquals(50L, ((BigDecimal) results.get(4)[1]).longValue());
     }
 }
